chore: apply new line-length

+++ src/aap_eda/core/migrations/0002_initial_data.py
@@ -32,9 +32,7 @@ def insert_permissions(apps, schema_editor):
     permissions = []
     for resource_type, actions in PERMISSIONS.items():
         for action in actions:
-            permissions.append(
-                permission_model(resource_type=resource_type, action=action)
-            )
+            permissions.append(permission_model(resource_type=resource_type, action=action))
     permission_model.objects.using(db_alias).bulk_create(permissions)
 
 
@@ -42,9 +40,7 @@ def drop_permissions(apps, schema_editor):
     permission_model = apps.get_model("core", "Permission")  # noqa: N806
     db_alias = schema_editor.connection.alias
     for resource_type, actions in PERMISSIONS.items():
-        permission_model.objects.using(db_alias).filter(
-            resource_type=resource_type, actions__in=actions
-        ).delete()
+        permission_model.objects.using(db_alias).filter(resource_type=resource_type, actions__in=actions).delete()
 
 
 class Migration(migrations.Migration):
